+++ packages/core-js/features/map/index.js
@@ -2,9 +2,12 @@ module.exports = require('../../es/map');
 
 require('../../modules/esnext.map.from');
 require('../../modules/esnext.map.of');
+require('../../modules/esnext.map.every');
 require('../../modules/esnext.map.filter');
 require('../../modules/esnext.map.group-by');
+require('../../modules/esnext.map.includes');
 require('../../modules/esnext.map.key-by');
 require('../../modules/esnext.map.map-keys');
 require('../../modules/esnext.map.map-values');
 require('../../modules/esnext.map.merge');
+require('../../modules/esnext.map.some');