+++ packages/core-js/features/map/some.js
@@ -1,0 +1,4 @@
+require('../../modules/es.map');
+require('../../modules/esnext.map.some');
+
+module.exports = require('../../internals/entry-unbind')('Map', 'some');