+++ packages/core-js/features/math/seeded-prng.js
@@ -1,0 +1,3 @@
+require('../../modules/esnext.math.seeded-prng');
+
+module.exports = require('../../internals/path').Math.seededPRNG;