+++ packages/core-js/features/symbol/dispose.js
@@ -1,0 +1,3 @@
+require('../../modules/esnext.symbol.dispose');
+
+module.exports = require('../../internals/wrapped-well-known-symbol').f('dispose');