+++ packages/core-js/modules/esnext.set.add-all.js
@@ -1,16 +1,10 @@
 'use strict';
-var anObject = require('../internals/an-object');
-var aFunction = require('../internals/a-function');
+var collectionAddAll = require('../internals/collection-add-all');
 
 // `Set.prototype.addAll` method
 // https://github.com/tc39/collection-methods
 require('../internals/export')({ target: 'Set', proto: true, real: true, forced: require('../internals/is-pure') }, {
   addAll: function addAll(/* ...elements */) {
-    var set = anObject(this);
-    var adder = aFunction(set.add);
-    for (var k = 0, len = arguments.length; k < len; k++) {
-      adder.call(set, arguments[k]);
-    }
-    return set;
+    return collectionAddAll.apply(this, arguments);
   }
 });