+++ packages/core-js/modules/esnext.symbol.dispose.js
@@ -1,0 +1,3 @@
+// `Symbol.patternMatch` well-known symbol
+// https://github.com/tc39/proposal-using-statement
+require('../internals/define-well-known-symbol')('dispose');