+++ packages/core-js/proposals/collection-methods.js
@@ -1,9 +1,12 @@
 require('../modules/esnext.map.group-by');
 require('../modules/esnext.map.key-by');
+require('../modules/esnext.map.every');
 require('../modules/esnext.map.filter');
+require('../modules/esnext.map.includes');
 require('../modules/esnext.map.map-keys');
 require('../modules/esnext.map.map-values');
 require('../modules/esnext.map.merge');
+require('../modules/esnext.map.some');
 require('../modules/esnext.set.add-all');
 require('../modules/esnext.set.delete-all');
 require('../modules/esnext.set.every');
@@ -13,3 +16,5 @@ require('../modules/esnext.set.join');
 require('../modules/esnext.set.map');
 require('../modules/esnext.set.reduce');
 require('../modules/esnext.set.some');
+require('../modules/esnext.weak-set.add-all');
+require('../modules/esnext.weak-set.delete-all');