+++ packages/core-js/stage/1.js
@@ -5,10 +5,12 @@ require('../proposals/math-extensions');
 require('../proposals/math-signbit');
 require('../proposals/number-from-string');
 require('../proposals/observable');
+require('../proposals/pattern-matching');
 require('../proposals/promise-try');
+require('../proposals/richer-keys');
+require('../proposals/seeded-random');
 require('../proposals/string-code-points');
 require('../proposals/string-replace-all');
-require('../proposals/richer-keys');
-require('../proposals/pattern-matching');
+require('../proposals/using-statement');
 
 module.exports = require('./2');