+++ packages/core-js/stage/2.js
@@ -1,4 +1,3 @@
-require('../proposals/object-from-entries');
 require('../proposals/set-methods');
 
 module.exports = require('./3');