+++ packages/core-js/stage/3.js
@@ -1,5 +1,6 @@
 require('../proposals/array-flat-and-flat-map');
 require('../proposals/global');
+require('../proposals/object-from-entries');
 require('../proposals/string-match-all');
 require('../proposals/string-trim-start-end');
 require('../proposals/symbol-description');